requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.0
PyPDF2==3.0.1
Pillow==10.1.0
matplotlib==3.8.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1